feat:  use pino logger and restructure helpers
