--- src/exampleFunction.test.ts
@@ -1,9 +1,0 @@
-import { describe, expect, it } from "vitest";
-import { exampleFunction } from "./exampleFunction.ts";
-
-describe("exampleFunction", () => {
-  it("should do something...", () => {
-    const result = exampleFunction("foo", "bar");
-    expect(result).toEqual(["Hello world!", "foo", "bar"]);
-  });
-});
