--- src/exampleFunction.ts
@@ -1,4 +1,0 @@
-export const exampleFunction = (...args: unknown[]) => {
-  console.log("Hello world!", args);
-  return ["Hello world!", ...args];
-};
