+++ src/helpers/logger/logger.ts
@@ -1,0 +1,3 @@
+import { pino } from "pino";
+
+export const logger = pino({ level: "debug" }, process.stdout);
