+++ src/server.ts
@@ -1,21 +1,18 @@
 import express from "express";
 import cors from "cors";
 import helmet from "helmet";
-import { config } from "./config.js";
-import { pinoHttp } from "pino-http";
-import { routes } from "./routes/index.js";
-import { logger } from "./helpers/logger.js";
 import "./services/supabase.js";
 import cookieParser from "cookie-parser";
-
-const { logger: pino } = pinoHttp();
-
-logger.debug("config", config);
+import { config } from "./config.ts";
+import { pinoHttp } from "pino-http";
+import { routes } from "./routes/index.ts";
+import { logger } from "./helpers/index.ts";
+import { corsOptions } from "./cors.ts";
 
 const checkConfigIsValid = () => {
   Object.values(config).forEach((value) => {
     if (!value) {
-      logger.error("config is invalid", config);
+      logger.error({ msg: "config is invalid", config });
       throw new Error("config is invalid");
     }
   });
@@ -36,7 +33,8 @@ app.use(cookieParser());
 
 app.use(
   pinoHttp({
-    logger: pino
+    logger: logger,
+    autoLogging: true
   })
 );
