+++ src/services/supabase.ts
@@ -1,4 +1,4 @@
-import { config } from "../config.js";
+import { config } from "../config.ts";
 import { createClient } from "@supabase/supabase-js";
 
 export const supabase = createClient(config.supabaseUrl, config.supabaseKey);
